refactor: move util class to main package
