+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/DatabaseClientImpl.java
@@ -56,10 +56,6 @@ class DatabaseClientImpl implements DatabaseClient {
     return pool.getSession();
   }
 
-  Session getRandomSession() {
-    return pool.getRandomSession();
-  }
-
   @Override
   public Dialect getDialect() {
     return pool.getDialect();
