+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/DatabaseClientImplTest.java
@@ -101,10 +101,10 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Base64;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Random;
+import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -1748,7 +1748,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
     try (ResultSet rs = client.singleUse().executeQuery(SELECT1)) {
       assertThat(rs.next()).isTrue();
@@ -3324,7 +3324,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
 
     // Getting a single use read-only transaction and not using it should not cause any sessions
@@ -3339,7 +3339,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
 
     client.singleUseReadOnlyTransaction().close();
@@ -3352,7 +3352,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
 
     client.readWriteTransaction();
@@ -3365,7 +3365,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
 
     client.readOnlyTransaction().close();
@@ -3378,7 +3378,7 @@ public class DatabaseClientImplTest {
     DatabaseClientImpl client =
         (DatabaseClientImpl)
             spanner.getDatabaseClient(DatabaseId.of(TEST_PROJECT, TEST_INSTANCE, TEST_DATABASE));
-    Collection<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
+    Set<PooledSessionFuture> checkedOut = client.pool.checkedOutSessions;
     assertThat(checkedOut).isEmpty();
 
     client.transactionManager().close();
